requirements.txt
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.1.8
dotenv==0.9.9
exceptiongroup==1.3.1
Flask==3.1.3
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector==2.2.9
mysql-connector-python==9.4.0
packaging==26.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.2
python-dotenv==1.2.1
tomli==2.4.0
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
zipp==3.23.0
